Created shooter class and updated collect class
